uploader_app/requirements.txt
Flask>=3.0.0
google-cloud-storage
google-cloud-firestore
Werkzeug>=3.0.0
gunicorn
